Added gameObjects ArrayList to store all constructed GameObjects in the game. Fixed various minor code snippets

+++ tankgame/util/Vector2D.java
@@ -12,7 +12,7 @@ public class Vector2D {
         this.y = 0;
     }
 
-    public Vector2D(int xValue, int yValue) {
+    public Vector2D(float xValue, float yValue) {
         this.x = xValue;
         this.y = yValue;
     }
